third_party/blink/web_tests/external/wpt/css/css-overflow/column-scroll-marker-focus-003-expected.txt

This is a testharness.js-based test.
[FAIL] Focus second column
  assert_equals: expected Element node <span tabindex="0" id="inlineElm1">\n      inline<br>\n    ... but got Element node <div tabindex="0" style="display:inline-block; width:100%...
[FAIL] Focus third column
  assert_equals: expected Element node <span tabindex="0" id="inlineElm2">inline2</span> but got Element node <div tabindex="0" style="display:inline-block; width:100%...
[FAIL] Focus fourth column
  assert_equals: expected Element node <div tabindex="0" id="inlineBlock1" style="display:inline... but got Element node <div tabindex="0" style="display:inline-block; width:100%...
[FAIL] Focus fifth column
  assert_equals: expected Element node <span tabindex="0" id="inlineElm3" style="border:1px soli... but got Element node <div tabindex="0" style="display:inline-block; width:100%...
[FAIL] Focus sixth column
  assert_equals: expected Element node <span tabindex="0" id="inlineElm4" style="border:1px soli... but got Element node <div tabindex="0" style="display:inline-block; width:100%...
Harness: the test ran to completion.
